==> 02_KNN model.ipynb <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python Docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the read-only "../input/" directory
# For example, running this (by clicking run or pressing Shift+Enter) will list all files under the input directory

import os
for dirname, _, filenames in os.walk('/kaggle/input'):
    for filename in filenames:
        print(os.path.join(dirname, filename))

# You can write up to 20GB to the current directory (/kaggle/working/) that gets preserved as output when you create a version using "Save & Run All" 
# You can also write temporary files to /kaggle/temp/, but they won't be saved outside of the current session

from surprise import Dataset, Reader, KNNBasic, accuracy
from surprise.model_selection import train_test_split, GridSearchCV
import pickle

import pandas as pd
import numpy as np
import time
df_pivot = pd.read_csv('/kaggle/input/naucno-izracunavanje/pivot_table.csv', index_col=0)
print(df_pivot.columns)

# Transformacija podataka u format pogodan za treniranje i učitavanje istih u Dataset objekat
df_melt = df_pivot.stack().reset_index().rename(columns={'level_1': 'Movie_Id', 0: 'Rating'})
print(df_melt.columns)
reader = Reader(rating_scale=(1, 5))
data = Dataset.load_from_df(df_melt[['Cust_Id', 'Movie_Id', 'Rating']], reader)

# Podela podataka na trening i validacioni skup
trainset, validset = train_test_split(data, test_size=0.25)

# Podešavanje parametara na kosinusnu slicnos i Item Based KNN
similarity_options = {'name': 'cosine', 'user_based': False}
best_rmse = float('inf')
best_model = None

# Trazimo najbolje k za kosinusnu slicnost (Radili smo rucno a ne grid search jer smo imali problema sa RAM-om na kaggle)
for k in [25, 30, 35, 40, 45]:
    algo = KNNBasic(k=k, sim_options=similarity_options)
    start_time = time.time()
    algo.fit(trainset)
    end_time = time.time()
    print(f"Vreme fitovanja: {end_time - start_time:.2f} sekundi")
    start_time = time.time()
    predictions = algo.test(validset)
    end_time = time.time()
    print(f"Vreme predvidjanja: {end_time - start_time:.2f} sekundi")
    rmse = accuracy.rmse(predictions, verbose=False)
    print(f'Cosine similarity, k={k}, RMSE={rmse}')
    
    if rmse < best_rmse:
        best_rmse = rmse
        best_model = algo
        best_params = {'similarity': 'cosine', 'k': k}


# Pronalazenje najbolje vrednosti za parametar k za Pearson korelaciju
similarity_options['name'] = 'pearson'
for k in [10, 15, 18, 20, 25]:
    algo = KNNBasic(k=k, sim_options=similarity_options)
    start_time = time.time()
    algo.fit(trainset)
    end_time = time.time()
    print(f"Vreme fitovanja: {end_time - start_time:.2f} sekundi")
    start_time = time.time()
    predictions = algo.test(validset)
    end_time = time.time()
    print(f"Vreme predvidjanja: {end_time - start_time:.2f} sekundi")
    rmse = accuracy.rmse(predictions, verbose=False)
    print(f'Pearson similarity, k={k}, RMSE={rmse}')
    
    if rmse < best_rmse:
        best_rmse = rmse
        best_model = algo
        best_params = {'similarity': 'pearson', 'k': k}

# Čuvanje modela
with open('best_knn_model.pkl', 'wb') as f:
    pickle.dump(best_model, f)

print(f'Najbolji model koristi {best_params["similarity"]} sličnost sa k={best_params["k"]} i RMSE={best_rmse}')


# Testiranje modela na test skupu
test_data = pd.read_csv('/kaggle/input/naucno-izracunavanje/test_data.csv')
test_data = Dataset.load_from_df(test_data[['Cust_Id', 'Movie_Id', 'Rating']], reader)
testset = test_data.build_full_trainset().build_testset()

with open('best_knn_model.pkl', 'rb') as f:
    best_model = pickle.load(f)

predictions = best_model.test(testset)
final_rmse = accuracy.rmse(predictions)

# Najbolji model smo dobili za Pearson korelaciju i k=18. Na trening skupu RMSE=0.9258 a na test RMSE=1.0623
