==> 07_Feature Engineering for Hybrid Model.ipynb <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python Docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the read-only "../input/" directory
# For example, running this (by clicking run or pressing Shift+Enter) will list all files under the input directory

import os
for dirname, _, filenames in os.walk('/kaggle/input'):
    for filename in filenames:
        print(os.path.join(dirname, filename))

# You can write up to 20GB to the current directory (/kaggle/working/) that gets preserved as output when you create a version using "Save & Run All" 
# You can also write temporary files to /kaggle/temp/, but they won't be saved outside of the current session

import pickle
import pandas as pd
from sklearn.model_selection import train_test_split
import numpy as np
import time

# Učitavanje KNN i SVD modela kako bi dobili njihova predviđanja koja ćemo kasnije koristiti kao featuree u našim hibridnim modelima
with open('/kaggle/input/netflix-models/best_knn_model.pkl', 'rb') as f:
    best_knn_model = pickle.load(f)

with open('/kaggle/input/netflix-models/best_svd_model.pkl', 'rb') as f:
    best_svd_model = pickle.load(f)

# Učitavanje trening skupa i transformacija u pogodan format
df_pivot = pd.read_csv('/kaggle/input/naucno-izracunavanje/pivot_table.csv', index_col=0)
print(df_pivot.head())

df_melt = df_pivot.stack().reset_index().rename(columns={'level_1': 'Movie_Id', 0: 'Rating'})
print(df_melt.head())

# Podela skupa na trening i validacioni
train_data, val_data = train_test_split(df_melt, test_size=0.25, random_state=42)

print(f"Trening podaci: {len(train_data)} redova")
print(f"Validacioni podaci: {len(val_data)} redova")

# Funkcija za generisanje dodatnih atributa koje ce koristiti hibridni modeli
# Funkcija generise: 
#     knn_pred - Predvidjanje KNN modela
#     svd_pred - Predvidjanje SVD modela
#     movie_mean_rating - Prosecna ocena filma
#     user_mean_rating - Prosecna ocena koju korisnik daje
#     user_review_count - Ukupan broj ocena koje je korisnik dao
#     movie_review_count - Ukupan broj ocena koje film ima
#     user_rating_variability - Varijansa ocena korisnika
    
def generate_features_targets(data, knn_model, svd_model, pivot_table):
    df_pivot = pivot_table.copy()
    
    knn_preds = np.array([knn_model.predict(row['Cust_Id'], row['Movie_Id']).est for _, row in data.iterrows()])
    svd_preds = np.array([svd_model.predict(row['Cust_Id'], row['Movie_Id']).est for _, row in data.iterrows()])
    
    movie_mean_ratings = df_pivot.mean()
    user_mean_ratings = df_pivot.mean(axis=1)
    user_review_counts = df_pivot.notna().sum(axis=1)
    movie_review_counts = df_pivot.notna().sum(axis=0)
    user_rating_variabilities = df_pivot.std(axis=1)
    
    data['knn_pred'] = data.apply(lambda row: knn_model.predict(row['Cust_Id'], row['Movie_Id']).est, axis=1)
    data['svd_pred'] = data.apply(lambda row: svd_model.predict(row['Cust_Id'], row['Movie_Id']).est, axis=1)
    data['movie_mean_rating'] = data['Movie_Id'].map(movie_mean_ratings)
    data['user_mean_rating'] = data['Cust_Id'].map(user_mean_ratings)
    data['user_review_count'] = data['Cust_Id'].map(user_review_counts)
    data['movie_review_count'] = data['Movie_Id'].map(movie_review_counts)
    data['user_rating_variability'] = data['Cust_Id'].map(user_rating_variabilities)
    
    features = data[['knn_pred', 'svd_pred', 'movie_mean_rating', 'user_mean_rating', 'user_review_count', 'movie_review_count', 'user_rating_variability']]
    targets = data['Rating']
    
    return features, targets

# Generišemo novi trening skup (skup sa novim atributima koji ćemo koristiti u hibridnim modelima)
start_time = time.time()
X_train, y_train = generate_features_targets(train_data, best_knn_model, best_svd_model, df_pivot)
end_time = time.time()
print(f"Vreme generisanja feature-a: {end_time - start_time:.2f} sekundi")

# Generišemo validacioni skup
start_time = time.time()
X_val, y_val = generate_features_targets(val_data, best_knn_model, best_svd_model, df_pivot)
end_time = time.time()
print(f"Vreme generisanja feature-a: {end_time - start_time:.2f} sekundi")

# Učitavanje starog test skupa
test_data = pd.read_csv('/kaggle/input/naucno-izracunavanje/test_data.csv')
print(test_data.head())

# Generišemo novi test skup transformacijom starog
start_time = time.time()
X_test, y_test = generate_features_targets(test_data, best_knn_model, best_svd_model, df_pivot)
end_time = time.time()
print(f"Vreme generisanja feature-a: {end_time - start_time:.2f} sekundi")

# Dodajemo ciljnu promenljivu našim podacima i čuvamo ih
X_train.loc[:, 'target'] = y_train
X_val.loc[:, 'target'] = y_val
X_test.loc[:, 'target'] = y_test

X_train.to_csv('train_features.csv', index=False)
X_val.to_csv('val_features.csv', index=False)
X_test.to_csv('test_features.csv', index=False)

import shutil

shutil.make_archive('/kaggle/working/features', 'zip', '/kaggle/working', 
                    'train_features.csv')
shutil.make_archive('/kaggle/working/features', 'zip', '/kaggle/working', 
                    'test_features.csv')
shutil.make_archive('/kaggle/working/features', 'zip', '/kaggle/working', 
                    'val_features.csv')

!cp /kaggle/working/features.zip /kaggle/input/

for dirname, _, filenames in os.walk('/kaggle/working'):
    for filename in filenames:
        print(os.path.join(dirname, filename))
